axonos_gate/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
websockets>=12.0
